Update output + figures
Added the main result from Hofman, Nelemans 2024 to the output folder. I
also updated the figures, which now display the LISA PI curve, generated
with Schnell, instead of the parabolic approximation as well as an
estimate of the galactic foreground based on Karnesis et al.
I also updated the README to update the explanation of the
normalisation.